backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
python-dotenv==1.0.0
pymongo==4.6.1
twilio==8.11.0
sendgrid==6.11.0
gunicorn==21.2.0
boto3==1.34.19
